separaioned bot.js for logic, one utils file
